Implement planning in drifty Odometry.

+++ utils/build_vimaps.sh
@@ -1,16 +1,16 @@
 #!/bin/bash
 
 # activate the maplab ROS environment
-source /home/michbaum/maplab_ws/devel/setup.bash
+source /home/michbaum/Projects/maplab/devel/setup.bash
 
 # Define the path to the directory containing your rosbag files
-rosbag_dir="/home/michbaum/optag_ws/data/sim_bags/query_trajectory_split"
+rosbag_dir="/home/michbaum/Projects/optag/data/sim_bags/warehouse_slowdown_gt_split"
 
 # Define the package containing your launch file
 package="maplab_node"
 
 # Define the name of your launch file
-launch_file="maze-maplab-node-w-rovioli.launch"
+launch_file="rosbag-maplab-node-w-rovioli.launch"
 
 # Change to the rosbag directory
 cd "$rosbag_dir"
@@ -20,16 +20,24 @@ rosbag_files=(*.bag)
 
 # Loop through each rosbag file
 for rosbag_file in "${rosbag_files[@]}"; do
-    echo "Playing rosbag: $rosbag_file"
-    
-    # Start the launch file in the background
-    roslaunch "$package" "$launch_file" &
-    launch_pid=$!
-    
-    # Start playing the rosbag
-    rosbag play "$rosbag_file"
-    
-    # Stop the launch file
-    kill -INT $launch_pid
-    wait $launch_pid
+    # Extract the number from the filename using awk
+    number=$(echo "$rosbag_file" | awk -F '[^0-9]*' '$0=$2')
+
+    # Check if the number is greater than 174
+    if (( number > 174 )); then
+        echo "Playing rosbag: $rosbag_file"
+        
+        # Start the launch file in the background
+        roslaunch "$package" "$launch_file" &
+        launch_pid=$!
+        
+        # Start playing the rosbag
+        rosbag play "$rosbag_file"
+        
+        # Stop the launch file
+        kill -INT $launch_pid
+        wait $launch_pid
+    else
+        echo "Skipping rosbag: $rosbag_file (Number is not greater than 174)"
+    fi
 done
